trimming model, trimming tests

+++ Cute Video Editor.Core/Models/CropFrameEntryModel.cs
@@ -1,8 +1,8 @@
 ﻿namespace CuteVideoEditor.Core.Models;
 
-public readonly struct CropFrameEntryModel(int frameNumber, RectModel rect)
+public readonly struct CropFrameEntryModel(long frameNumber, RectModel rect)
 {
-    public int FrameNumber { get; } = frameNumber;
+    public long FrameNumber { get; } = frameNumber;
     public RectModel Rect { get; } = rect;
 }
